requirements.txt
absl-py==0.14.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Limiter==1.4
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
limits==1.5.1
Markdown==3.3.4
MarkupSafe==2.0.1
nltk==3.6.3
numpy==1.19.2
oauthlib==3.1.1
opt-einsum==3.3.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.12.0
python-dotenv==0.19.0
regex==2021.9.24
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.1
wrapt==1.12.1
